refctor: location test
